Rename: reservation 폴더로 예매 관련 기능 이동
