+++ src/main/java/org/baseball/domain/reservation/ReservationMapper.java
@@ -1,0 +1,5 @@
+package org.baseball.domain.reservation;
+
+public interface ReservationMapper {
+}
+
