+++ src/main/java/org/baseball/domain/reservation/ReservationService.java
@@ -1,0 +1,4 @@
+package org.baseball.domain.reservation;
+
+public class ReservationService {
+}
