+++ src/main/webapp/WEB-INF/views/reservation/tickets2.jsp
@@ -2,9 +2,9 @@
 <html>
 <head>
     <title></title>
-    <link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/tickets/tickets-common.css">
-    <link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/tickets/tickets-common2.css">
-    <link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/tickets/tickets2.css">
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/reservation/tickets-common.css">
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/reservation/tickets-common2.css">
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/reservation/tickets2.css">
 </head>
 <body>
 <!-- 권종/할인 선택 -->
@@ -244,14 +244,14 @@
                 </div>
                 <div class="ticket-btn-container">
                     <div class="ticket-btn-wrapper">
-                        <img src="${pageContext.request.contextPath}/assets/img/tickets/backBtnSmall.svg" alt="이전버튼">
-                        <img src="${pageContext.request.contextPath}/assets/img/tickets/nextBtnSmall.svg" alt="다음버튼">
+                        <img src="${pageContext.request.contextPath}/assets/img/reservation/backBtnSmall.svg" alt="이전버튼">
+                        <img src="${pageContext.request.contextPath}/assets/img/reservation/nextBtnSmall.svg" alt="다음버튼">
                     </div>
                 </div>
             </div>
         </div>
     </div>
 </div>
-<script type="module" src="${pageContext.request.contextPath}/assets/js/tickets/tickets2.js"></script>
+<script type="module" src="${pageContext.request.contextPath}/assets/js/reservation/tickets2.js"></script>
 </body>
 </html>
