+++ src/main/webapp/WEB-INF/views/reservation/toolbar.jsp
@@ -1,0 +1,16 @@
+<link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/reservation/toolbar.css" />
+<div class="btn-container">
+    <div class="btn-wrapper" id="gobackBtn">
+        <img src="${pageContext.request.contextPath}/assets/img/reservation/gobackBtn.svg" alt="뒤로가기 버튼">
+    </div>
+    <div class="btn-wrapper" id="enlargementBtn">
+        <img src="${pageContext.request.contextPath}/assets/img/reservation/enlargementBtn.svg" alt="확대 버튼">
+    </div>
+    <div class="btn-wrapper" id="reductionBtn">
+        <img src="${pageContext.request.contextPath}/assets/img/reservation/reductionBtn.svg" alt="축소 버튼">
+    </div>
+    <div class="btn-wrapper" id="reloadBtn">
+        <img src="${pageContext.request.contextPath}/assets/img/reservation/reloadBtn.svg" alt="새로고침 버튼">
+    </div>
+</div>
+<script type="module" src="${pageContext.request.contextPath}/assets/js/reservation/toolbar.js"></script>
