--- src/main/webapp/WEB-INF/views/tickets/toolbar.jsp
@@ -1,16 +1,0 @@
-<link rel="stylesheet" href="${pageContext.request.contextPath}/assets/css/tickets/toolbar.css" />
-<div class="btn-container">
-    <div class="btn-wrapper" id="gobackBtn">
-        <img src="${pageContext.request.contextPath}/assets/img/tickets/gobackBtn.svg" alt="뒤로가기 버튼">
-    </div>
-    <div class="btn-wrapper" id="enlargementBtn">
-        <img src="${pageContext.request.contextPath}/assets/img/tickets/enlargementBtn.svg" alt="확대 버튼">
-    </div>
-    <div class="btn-wrapper" id="reductionBtn">
-        <img src="${pageContext.request.contextPath}/assets/img/tickets/reductionBtn.svg" alt="축소 버튼">
-    </div>
-    <div class="btn-wrapper" id="reloadBtn">
-        <img src="${pageContext.request.contextPath}/assets/img/tickets/reloadBtn.svg" alt="새로고침 버튼">
-    </div>
-</div>
-<script type="module" src="${pageContext.request.contextPath}/assets/js/tickets/toolbar.js"></script>
