+++ src/main/webapp/assets/css/reservation/tickets-common.css
