+++ src/main/webapp/assets/css/reservation/tickets3.css
@@ -347,14 +347,14 @@
 
 .orderConfirm-checkbox+label {
     background-repeat: no-repeat;
-    background-image: url('/assets/img/tickets/checkboxOff.svg');
+    background-image: url('/assets/img/reservation/checkboxOff.svg');
     background-position: center center;
     background-size: contain;
 }
 
 .orderConfirm-checkbox:checked+label {
     background-repeat: no-repeat;
-    background-image: url('/assets/img/tickets/checkboxOn.svg');
+    background-image: url('/assets/img/reservation/checkboxOn.svg');
     background-position: center center;
     background-size: contain;
 }
