+++ src/main/webapp/assets/js/reservation/seats.js
